Modified locking mechanism to enable higher streaming performance to DRAM.
Previously, a cache hit to DRAM would lock an entire page, which would serialize
subsequent accesses to the same page. The goal of this patch is to allow accesses
to different physical addresses within the same page to go to DRAMSim in parallel,
allowing for higher throughput. It adds the pending_flash_addr map to track
the accesses at a finer granularity and only locks the entire page on a cache miss.