Tracking of unused prefetches.
This change adds a prefetched and used flag to the cache_line structure.
These flags are set appropriately to track whether a cache line was
brought into the cache by a prefetch operation and whether or not the
line was later used by a non-prefetch operation.

The new cache_line flags are then used to compute the counters
unused_prefetches (the number of prefetched pages n the cace that have
not yet been used) and unused_prefetch_victims (the number of prefetches
that were never used before being evicted from the cache).  This should
help us understand the efficiency of our prefetchers better in the future.