Start of stream buffer implementation.
Created the state necessary for the stream buffer and implemented
the stream_buffer_miss_handler(), which handles recognizing streams
and issuing the initial prefetches to start a stream buffer.

The stream_buffer_hit_handler() still needs to be implemented to
handle further prefetching in the stream. This also needs to be
tested heavily before merging back into the dev branch.