More updates to NVDIMM ini files.
From now on, the NVDIMMSim repo located at
https://github.com/jimstevens2001/NVDIMMSim is the only "officially"
supported NVDIMMSim repo. The stable branch there is guaranteed to work
(or at least will be in the near future) with the dev and master branch
of HybridSim.  Tagged releases of NVDIMMSim will also be provided to match
the tagged releases of HybridSim and DRAMSim. Otherwise, all bets are
off. The DRAMSim and NVDIMM branches ran by Paul Rosenfeld (DRAMNinjas)
and Paul Tschirhart (watchmaker) are not necessarily compatible with
HybridSim. If you need a feature they provide, please let me know and
I can pull it into my copy of the repos. --Jim