Added C interface for Python frontend.
The C interface enables users of HybridSim to create instances of
HybridSim using Python's ctypes.cdll interface.  The only complication
here is that cdll cannot handle callbacks, so this interface requires
the user to poll.  A commit in the near future will provide a Python
class to wrap the C interface.