HybridSim Python Interface
Using the C external interface and Python's ctypes module, the
hybridsim.py file creates a Python wrapper to HybridSim. The main function
of the file shows how to instantiate and use it.  This will be used as
the basis for a Python version of the trace-based simulation and later
for a specialized prefetching simulator.