Add support for base addresses in MT_TBS
This is useful to allow MT_TBS to reuse the same trace for
different threads. It works by just creating a different base
address for each thread and adding that to the requested address.

In the future, we might also add a mode to take into account kernel
regions, so those addresses won't be adjusted.